+++ feature/widget/src/main/java/com/onmi/widget/combined/CombinedInfo.kt
@@ -1,0 +1,4 @@
+package com.onmi.widget.combined
+
+interface CombinedInfo {
+}
