+++ feature/widget/src/main/java/com/onmi/widget/combined/CombinedWorker.kt
@@ -1,0 +1,4 @@
+package com.onmi.widget.combined
+
+class CombinedWorker {
+}
