--- feature/widget/src/main/java/com/onmi/widget/glance/combined/CombinedInfo.kt
@@ -1,4 +1,0 @@
-package com.onmi.widget.glance.combined
-
-interface CombinedInfo {
-}
