--- feature/widget/src/main/java/com/onmi/widget/glance/combined/CombinedWorker.kt
@@ -1,4 +1,0 @@
-package com.onmi.widget.glance.combined
-
-class CombinedWorker {
-}
