+++ feature/widget/src/main/java/com/onmi/widget/theme/ONMIWidgetColorScheme.kt
@@ -1,4 +1,4 @@
-package com.onmi.widget.glance.theme
+package com.onmi.widget.theme
 
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
