+++ feature/widget/src/main/java/com/onmi/widget/timetable/TimeTableWidget.kt
@@ -1,4 +1,4 @@
-package com.onmi.widget.glance.meal
+package com.onmi.widget.timetable
 
 import android.content.Context
 import android.os.Build
@@ -20,37 +20,36 @@ import androidx.glance.background
 import androidx.glance.currentState
 import androidx.glance.layout.Alignment
 import androidx.glance.layout.Column
+import androidx.glance.layout.Row
 import androidx.glance.layout.Spacer
 import androidx.glance.layout.fillMaxSize
 import androidx.glance.layout.height
 import androidx.glance.layout.padding
-import com.onmi.widget.glance.theme.ONMIWidgetColorScheme
-import com.onmi.widget.glance.util.SuitText
+import androidx.glance.layout.width
+import com.onmi.widget.theme.ONMIWidgetColorScheme
+import com.onmi.widget.util.SuitText
 
-class MealWidget : GlanceAppWidget() {
-    override val stateDefinition = MealInfoStateDefinition
+class TimeTableWidget : GlanceAppWidget() {
+    override val stateDefinition = TimeTableInfoStateDefinition
 
     @RequiresApi(Build.VERSION_CODES.O)
     override suspend fun provideGlance(context: Context, id: GlanceId) {
         provideContent {
             LaunchedEffect(key1 = Unit) {
-                MealWorker.enqueue(context)
+                TimeTableWorker.enqueue(context)
             }
 
             GlanceTheme(colors = ONMIWidgetColorScheme.colors) {
-                when (val state = currentState<MealInfo>()) {
-                    is MealInfo.Available -> {
-                        SuccessContent(
-                            mealTime = state.mealTime,
-                            mealList = state.mealList
-                        )
+                when (val state = currentState<TimeTableInfo>()) {
+                    is TimeTableInfo.Available -> {
+                        SuccessContent(timeTable = state.timeTableData)
                     }
 
-                    is MealInfo.Loading -> {
+                    is TimeTableInfo.Loading -> {
                         LoadingContent()
                     }
 
-                    is MealInfo.Unavailable -> {
+                    is TimeTableInfo.Unavailable -> {
                         UnavailableContent()
                     }
                 }
@@ -59,10 +58,7 @@ class MealWidget : GlanceAppWidget() {
     }
 
     @Composable
-    private fun SuccessContent(
-        mealTime: String,
-        mealList: List<String>,
-    ) {
+    private fun SuccessContent(timeTable: List<String>) {
         val context = LocalContext.current
 
         LazyColumn(
@@ -71,25 +67,24 @@ class MealWidget : GlanceAppWidget() {
                 .padding(12.dp)
                 .background(GlanceTheme.colors.onPrimary),
         ) {
-            item {
-                SuitText(
-                    text = "[$mealTime]",
-                    color = GlanceTheme.colors.primary.getColor(context),
-                    fontSize = 14.sp
-                )
-                Spacer(modifier = GlanceModifier.height(1.37.dp))
-            }
-            itemsIndexed(mealList) { index, item ->
-                val text = if (item.length >= 14) item.substring(0, 13) + "..." else item
-
+            itemsIndexed(timeTable) { index, item ->
                 Column {
-                    SuitText(
-                        text = text,
-                        color = GlanceTheme.colors.tertiary.getColor(context),
-                        fontSize = 14.sp
-                    )
-                    if (mealList.lastIndex != index) {
-                        Spacer(modifier = GlanceModifier.height(1.37.dp))
+                    Row(verticalAlignment = Alignment.CenterVertically) {
+                        SuitText(
+                            modifier = GlanceModifier.width(12.dp),
+                            text = "${index + 1}",
+                            color = GlanceTheme.colors.secondary.getColor(context),
+                            fontSize = 14.sp,
+                        )
+                        Spacer(modifier = GlanceModifier.width(2.dp))
+                        SuitText(
+                            text = item,
+                            color = GlanceTheme.colors.tertiary.getColor(context),
+                            fontSize = 14.sp,
+                        )
+                    }
+                    if (timeTable.lastIndex != index) {
+                        Spacer(modifier = GlanceModifier.height(6.68.dp))
                     }
                 }
             }
@@ -108,7 +103,7 @@ class MealWidget : GlanceAppWidget() {
             verticalAlignment = Alignment.CenterVertically
         ) {
             SuitText(
-                text = "급식 정보를 불러올 수 없습니다.",
+                text = "시간표 정보를 불러올 수 없습니다.",
                 color = GlanceTheme.colors.tertiary.getColor(context),
                 fontSize = 14.sp,
             )
@@ -127,7 +122,7 @@ class MealWidget : GlanceAppWidget() {
             verticalAlignment = Alignment.CenterVertically
         ) {
             SuitText(
-                text = "급식 정보 불러오는중..",
+                text = "시간표 정보 불러오는중..",
                 color = GlanceTheme.colors.tertiary.getColor(context),
                 fontSize = 14.sp,
             )
@@ -135,7 +130,7 @@ class MealWidget : GlanceAppWidget() {
     }
 }
 
-class MealWidgetReceiver : GlanceAppWidgetReceiver() {
+class TimeTableWidgetReceiver : GlanceAppWidgetReceiver() {
     override val glanceAppWidget: GlanceAppWidget
-        get() = MealWidget()
+        get() = TimeTableWidget()
 }
