+++ feature/widget/src/main/java/com/onmi/widget/timetable/TimeTableWorker.kt
@@ -1,4 +1,4 @@
-package com.onmi.widget.glance.timetable
+package com.onmi.widget.timetable
 
 import android.content.Context
 import android.os.Build
